+++ avm/src/main.rs
@@ -2,6 +2,7 @@ use anyhow::{anyhow, Error, Result};
 use avm::InstallTarget;
 use clap::{CommandFactory, Parser, Subcommand};
 use semver::Version;
+use std::ffi::OsStr;
 
 #[derive(Parser)]
 #[clap(name = "avm", about = "Anchor version manager", version)]
@@ -19,9 +20,12 @@ pub enum Commands {
     },
     #[clap(about = "Install a version of Anchor", alias = "i")]
     Install {
-        /// Anchor version or commit
-        #[clap(value_parser = parse_install_target)]
-        version_or_commit: InstallTarget,
+        /// Anchor version or commit, conflicts with `--path`
+        #[clap(required_unless_present = "path")]
+        version_or_commit: Option<String>,
+        /// Path to local anchor repo, conflicts with `version_or_commit`
+        #[clap(long, conflicts_with = "version_or_commit")]
+        path: Option<String>,
         #[clap(long)]
         /// Flag to force installation even if the version
         /// is already installed
@@ -51,27 +55,21 @@ fn parse_version(version: &str) -> Result<Version, Error> {
     if version == "latest" {
         avm::get_latest_version()
     } else {
-        Version::parse(version).map_err(|e| anyhow!(e))
+        Ok(Version::parse(version)?)
     }
 }
 
 fn parse_install_target(version_or_commit: &str) -> Result<InstallTarget, Error> {
-    parse_version(version_or_commit)
-        .map(|version| {
-            if version.pre.is_empty() {
-                InstallTarget::Version(version)
-            } else {
-                // Allow `avm install 0.28.0-6cf200493a307c01487c7b492b4893e0d6f6cb23`
-                InstallTarget::Commit(version.pre.to_string())
-            }
-        })
-        .or_else(|version_error| {
-            avm::check_and_get_full_commit(version_or_commit)
-                .map(InstallTarget::Commit)
-                .map_err(|commit_error| {
-                    anyhow!("Not a valid version or commit: {version_error}, {commit_error}")
-                })
-        })
+    if let Ok(version) = parse_version(version_or_commit) {
+        if version.pre.is_empty() {
+            return Ok(InstallTarget::Version(version));
+        }
+        // Allow for e.g. `avm install 0.28.0-6cf200493a307c01487c7b492b4893e0d6f6cb23`
+        return Ok(InstallTarget::Commit(version.pre.to_string()));
+    }
+    avm::check_and_get_full_commit(version_or_commit)
+        .map(InstallTarget::Commit)
+        .map_err(|e| anyhow!("Not a valid version or commit: {e}"))
 }
 
 pub fn entry(opts: Cli) -> Result<()> {
@@ -79,9 +77,17 @@ pub fn entry(opts: Cli) -> Result<()> {
         Commands::Use { version } => avm::use_version(version),
         Commands::Install {
             version_or_commit,
+            path,
             force,
             from_source,
-        } => avm::install_version(version_or_commit, force, from_source),
+        } => {
+            let install_target = if let Some(path) = path {
+                InstallTarget::Path(path.into())
+            } else {
+                parse_install_target(&version_or_commit.unwrap())?
+            };
+            avm::install_version(install_target, force, from_source)
+        }
         Commands::Uninstall { version } => avm::uninstall_version(&version),
         Commands::List {} => avm::list_versions(),
         Commands::Update {} => avm::update(),
@@ -92,7 +98,54 @@ pub fn entry(opts: Cli) -> Result<()> {
     }
 }
 
+fn anchor_proxy() -> Result<()> {
+    let args = std::env::args().skip(1).collect::<Vec<String>>();
+
+    let version = avm::current_version()
+        .map_err(|_e| anyhow::anyhow!("Anchor version not set. Please run `avm use latest`."))?;
+
+    let binary_path = avm::version_binary_path(&version);
+    if !binary_path.exists() {
+        anyhow::bail!(
+            "anchor-cli {} not installed. Please run `avm use {}`.",
+            version,
+            version
+        );
+    }
+
+    let exit = std::process::Command::new(binary_path)
+        .args(args)
+        .env(
+            "PATH",
+            format!(
+                "{}:{}",
+                avm::get_bin_dir_path().to_string_lossy(),
+                std::env::var("PATH").unwrap_or_default()
+            ),
+        )
+        .spawn()?
+        .wait_with_output()
+        .expect("Failed to run anchor-cli");
+
+    if !exit.status.success() {
+        std::process::exit(exit.status.code().unwrap_or(1));
+    }
+
+    Ok(())
+}
+
 fn main() -> Result<()> {
+    // If the binary is named `anchor` then run the proxy.
+    if let Some(stem) = std::env::args()
+        .next()
+        .as_ref()
+        .and_then(|s| std::path::Path::new(s).file_stem().and_then(OsStr::to_str))
+    {
+        if stem == "anchor" {
+            return anchor_proxy();
+        }
+    }
+
     // Make sure the user's home directory is setup with the paths required by AVM.
     avm::ensure_paths();