+++ lang/syn/src/codegen/accounts/constraints.rs
@@ -360,12 +360,15 @@ pub fn generate_constraint_raw(ident: &Ident, c: &ConstraintRaw) -> proc_macro2:
 
 pub fn generate_constraint_owner(f: &Field, c: &ConstraintOwner) -> proc_macro2::TokenStream {
     let ident = &f.ident;
-    let maybe_deref = match &f.ty {
+    let maybe_deref = if match &f.ty {
         Ty::Account(AccountTy { boxed, .. })
         | Ty::InterfaceAccount(InterfaceAccountTy { boxed, .. }) => *boxed,
         _ => false,
-    }
-    .then_some(quote!(*));
+    } {
+        quote!(*)
+    } else {
+        Default::default()
+    };
     let owner_address = &c.owner_address;
     let error = generate_custom_error(
         ident,